Version 1.1
Added the base noise wrapper class for the FastNoise functionality.
Hasn't been tested yet but everything compiles fine and shows in the editor.
Made a lot of changes to mesh generation.
Different functions and constructors work for different uses e.g. editor, drag-and-drop so a lot of things were changed.
Currently everything happens in the OnConstruction method which generates a new mesh or updates an existing one.
An RMC custom provider has been made closely following the existing SphereProvider code. Will be edited to include noise at a later version.
Customization was attempted to be added to the noise classes for better in-editor usage but it wasn't implemented due to complexity and lack of access to enums. Instead they were hierarchically inheritted on a need-to-have basis. Customization class was left for any further changes that can be made.