Version 1.2
Tested out noise settings.
All types of noise work as intended.
Started the process of adding metadata tags to the noise class for better functionality.
Found a bug where setting frequency to 0.00f breaks the noise object and requires a reset of the whole layer. To fix this the value was set a minimum of 0.01 as debugging didnt have any definite results.
Removed white noise and crater noise types since crater didn't provide any special results while white noise created an obvious seam at the edge of the planet. It will be good to reimplement it at a later stage if there is the possibility of choosing an area affect by noise instead of having it everywhere.