Version 1.3
Changed RuntimeSphereGenerator class to ProceduralPlanetActor class, it has the same functionality but it fits better name-wise with the API.
Fixed issues with initialization errors during runtime and changed Noise pointers to a PlanetSettings pointer in all classes.

TODO for next version: Overhaul the provider to use the radius and resolution from inside the settings as it already has a pointer to it.