Version 1.4
Left the refactor of the provider for a later stage of development.
Added UI functionality for customizing the editor.
Now it is possible to spawn a singular blank or random planet from the editor.
Added a custom logo for the blank planket, have to add one for the random button as well at a later stage.
Added functionality to focus the camera on the newly created object upon planet creation.
Added a planet seed variable which controls all aspects of generation if the planet is randomized.
Added noise functions for creating random layers of each type of noise.
Added new UI styles and commands. Can try to Separate functionality outside of the main plugin class at a later version to increase decoupling.