Version 1.6 bug fixes
Slightly altered generation numbers for noise.
Made U and V texture tiling to be generated with the same number to prevent seams in texture transitions.
Added various validations in the provider and material.
Added a default texture if one is removed from the material, changed default values for texture transitions.
Added actor selection support for newly created planets to be immediately selected.
Found a bug where undo does not work on the planet objects.