Version 1.7
Made the Planet Settings properties for Material and Noise Settings private. All access to them is rerouted to the Planet Settings class.
Fixed a bug, where removing the material and setting it back to the procedural planet material, then trying to edit the properties would crash the editor.
Started debugging the not saving feature.